requirements.txt
python-dotenv
llama-index-core
llama-parse
llama-index-readers-file
llama-index-llms-openai
llama-index-embeddings-openai
